- correction

+++ src/main/java/ch/qa/testautomation/tas/common/IOUtils/FileOperation.java
@@ -205,11 +205,14 @@ public class FileOperation {
     }
 
     public static void deleteFolder(File folder) {
-        try (Stream<Path> paths = Files.walk(folder.toPath())) {
-            paths.sorted(Comparator.reverseOrder()).forEach(file -> FileOperation.deleteFile(file.toFile()));
-        } catch (IOException ex) {
-            throw new ExceptionBase(ExceptionErrorKeys.IOEXCEPTION_GENERAL, ex, "Failed on deleting folder: " + folder.getPath());
-        }
+        if (!folder.exists()) {
+            try (Stream<Path> paths = Files.walk(folder.toPath())) {
+                paths.sorted(Comparator.reverseOrder()).forEach(file -> FileOperation.deleteFile(file.toFile()));
+            } catch (IOException ex) {
+                throw new ExceptionBase(ExceptionErrorKeys.IOEXCEPTION_GENERAL, ex, "Failed on deleting folder: " + folder.getPath());
+            }
+        } else info("Delete folder: " + folder + " does not exist!");
+
     }
 
     /**
